Co-existing hashed and non-hashed password
This set of changes allows co-existing of both types
of keys in code. The next step would be to silently migrate
between the two.
Co-existing hashed and non-hashed password

This set of changes allows co-existing of both types
of keys in code. The next step would be to silently migrate
between the two.