Address timeout issue
When configured the database will lock after a certain period of time.
Every time that there is activity, the timeout counter is reset.
For example, if you configure the timeout to be 600 seconds, the
database will lock after 10 minutes of inactivity. If you show a
password after 9 minutes and 50 seconds, the database base will stay
open for another whole 10 minutes.

Issue #71